requirements.txt
aws-cdk-lib==2.104.0
boto3
constructs>=10.0.0,<11.0.0
pyjwt
requests
